navbar converted to server component, cvdownloader component added which is a client component to download cv
